requirements.txt
aiogram==2.23.1
fastapi==0.87.0
python-dotenv==0.21.0
uvicorn==0.19.0
psycopg2-binary
aioschedule~=0.5.2
openai~=0.26.1
transformers~=4.25.1
googletrans~=3.1.0a0
youtube-transcript-api
python-dateutil~=2.8.2
glqiwiapi~=2.18.3